requirements.txt
backports.zoneinfo==0.2.1
blis==0.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==8.1.7
cymem==2.0.8
dateparser==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#sha256=4db16860a8cdef56d436038ace6abeb9181a5176bdc8c16c755a20dced51e5f1
Flask==2.2.5
greenlet==3.0.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mathparse==0.1.2
murmurhash==1.0.10
nltk==3.8.1
numpy==1.21.6
peewee==3.17.0
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==3.13
regex==2023.10.3
requests==2.31.0
six==1.16.0
spacy==2.1.9
srsly==1.0.7
thinc==7.0.8
tqdm==4.66.1
typing_extensions==4.7.1
tzlocal==5.1
urllib3==2.0.7
wasabi==0.10.1
Werkzeug==2.2.3
zipp==3.15.0
